requirements.txt
functions-framework==3.*
pandas-gbq
pandas
gcsfs
fsspec
google-cloud-storage
